migrate namespace attachment controller actions to concern
