+++ app/controllers/groups/attachments_controller.rb
@@ -2,87 +2,24 @@
 
 module Groups
   # Controller actions for Group Attachments
-  class AttachmentsController < Groups::ApplicationController # rubocop:disable Metrics/ClassLength
+  class AttachmentsController < Groups::ApplicationController
     include Metadata
-    before_action :group, :current_page
-    before_action :new_destroy_params, only: %i[new_destroy]
-    before_action :attachment, only: %i[destroy]
+    include AttachmentActions
 
-    def index
-      authorize! @group, to: :view_attachments?
-
-      @render_individual_attachments = filter_requested?
-      @q = build_ransack_query
-      set_default_sort
-      @pagy, @attachments = pagy_with_metadata_sort(@q.result)
-    end
-
-    def new
-      authorize! @group, to: :create_attachment?
-
-      render turbo_stream: turbo_stream.update('attachment_modal',
-                                               partial: 'new_attachment_modal',
-                                               locals: {
-                                                 open: true,
-                                                 attachment: Attachment.new(attachable: @group),
-                                                 namespace: @group
-                                               }), status: :ok
-    end
-
-    def create
-      @attachments = ::Attachments::CreateService.new(current_user, @group, attachment_params).execute
-
-      status = if !@attachments.count.positive?
-                 :unprocessable_entity
-               elsif @attachments.count(&:persisted?) == @attachments.count
-                 :ok
-               else
-                 :multi_status
-               end
-
-      respond_to do |format|
-        format.turbo_stream do
-          render status:, locals: { attachment: Attachment.new(attachable: @group),
-                                    attachments: @attachments }
-        end
-      end
-    end
-
-    def new_destroy
-      authorize! @group, to: :destroy_attachment?
-      render turbo_stream: turbo_stream.update('attachment_modal',
-                                               partial: 'delete_attachment_modal',
-                                               locals: {
-                                                 open: true,
-                                                 attachment: @attachment,
-                                                 namespace: @group
-                                               }), status: :ok
-    end
-
-    def destroy # rubocop:disable Metrics/MethodLength
-      @destroyed_attachments = ::Attachments::DestroyService.new(@group, @attachment, current_user).execute
-      respond_to do |format|
-        if @destroyed_attachments.count.positive?
-          status = destroy_status(@attachment, @destroyed_attachments.length)
-          format.turbo_stream do
-            render status:, locals: { destroyed_attachments: @destroyed_attachments }
-          end
-        else
-          format.turbo_stream do
-            render status: :unprocessable_entity,
-                   locals: { message: t('.error',
-                                        filename: @attachment.file.filename,
-                                        errors: @attachment.errors.full_messages.first),
-                             destroyed_attachments: nil }
-          end
-        end
-      end
-    end
+    before_action :group
 
     private
 
     def group
       @group = Group.find_by_full_path(params[:group_id]) # rubocop:disable Rails/DynamicFindBy
+    end
+
+    def set_model
+      @namespace = group
+    end
+
+    def set_authorization_object
+      @authorize_object = group
     end
 
     def current_page
@@ -94,51 +31,8 @@ module Groups
       @context_crumbs +=
         [{
           name: t(:'groups.sidebar.files'),
-          path: group_attachments_path(@group)
+          path: group_attachments_path(@namespace)
         }]
-    end
-
-    def filter_requested?
-      params.dig(:q, :puid_or_file_blob_filename_cont).present?
-    end
-
-    def build_ransack_query
-      if @render_individual_attachments
-        @group.attachments.all.ransack(params[:q])
-      else
-        @group.attachments
-              .where.not(Attachment.arel_table[:metadata].contains({ direction: 'reverse' }))
-              .ransack(params[:q])
-      end
-    end
-
-    def layout_fixed
-      super
-      return unless action_name == 'index'
-
-      @fixed = false
-    end
-
-    def set_default_sort
-      @q.sorts = 'updated_at desc' if @q.sorts.empty?
-    end
-
-    def attachment_params
-      params.require(:attachment).permit(:attachable_id, :attachable_type, files: [])
-    end
-
-    def new_destroy_params
-      @attachment = Attachment.find_by(id: params[:attachment_id])
-    end
-
-    def attachment
-      @attachment = Attachment.find_by(id: params[:id])
-    end
-
-    def destroy_status(attachment, count)
-      return count == 2 ? :ok : :multi_status if attachment.associated_attachment
-
-      count == 1 ? :ok : :unprocessable_entity
     end
   end
 end
