+++ config/locales/en.yml
@@ -135,6 +135,9 @@ en:
         provider: Authentication Provider
         uid: Authentication Provider UID
     concerns:
+      atttachment_actions:
+        destroy:
+          error: 'File %{filename} was not removed due to the following errors: %{errors}'
       history:
         system: System
       track_activity:
@@ -585,7 +588,6 @@ en:
         failure: 'File %{filename} was not uploaded due to the following errors: %{errors}'
         success: File %{filename} was successfully uploaded.
       destroy:
-        error: 'File %{filename} was not removed due to the following errors: %{errors}'
         success: File %{filename} was successfully removed.
       index:
         search:
@@ -999,7 +1001,6 @@ en:
         failure: 'File %{filename} was not uploaded due to the following errors: %{errors}'
         success: File %{filename} was successfully uploaded.
       destroy:
-        error: 'File %{filename} was not removed due to the following errors: %{errors}'
         success: File %{filename} was successfully removed.
       index:
         search:
