+++ config/locales/fr.yml
@@ -135,6 +135,9 @@ fr:
         provider: Authentication Provider
         uid: Authentication Provider UID
     concerns:
+      attachment_actions:
+        destroy:
+          error: 'File %{filename} was not removed due to the following errors: %{errors}'
       history:
         system: System
       track_activity:
@@ -585,7 +588,6 @@ fr:
         failure: 'File %{filename} was not uploaded due to the following errors: %{errors}'
         success: File %{filename} was successfully uploaded.
       destroy:
-        error: 'File %{filename} was not removed due to the following errors: %{errors}'
         success: File %{filename} was successfully removed.
       index:
         search:
@@ -999,7 +1001,6 @@ fr:
         failure: 'File %{filename} was not uploaded due to the following errors: %{errors}'
         success: File %{filename} was successfully uploaded.
       destroy:
-        error: 'File %{filename} was not removed due to the following errors: %{errors}'
         success: File %{filename} was successfully removed.
       index:
         search:
